Cambio de short por Integer

+++ src/main/java/ec/edu/espe/arquitectura/escolastico/educacion/model/Calificacion.java
@@ -54,7 +54,7 @@ public class Calificacion implements Serializable {
         this.pk = calificacionPK;
     }
 
-    public Calificacion(String codMatricula, short codNrc) {
+    public Calificacion(String codMatricula, Integer codNrc) {
         this.pk = new CalificacionPK(codMatricula, codNrc);
     }
