+++ src/main/java/ec/edu/espe/arquitectura/escolastico/educacion/model/Nrc.java
@@ -31,9 +31,9 @@ public class Nrc implements Serializable {
     @EmbeddedId
     private NrcPK pk;
     @Column(name = "cupo_disponible", nullable = false)
-    private short cupoDisponible;
+    private Integer cupoDisponible;
     @Column(name = "cupo_registrado", nullable = false)
-    private short cupoRegistrado;
+    private Integer cupoRegistrado;
     @Column(name = "nombre", length = 255)
     private String nombre;
     @Column(name = "cod_persona", nullable = false, insertable = false, updatable = false)
@@ -71,19 +71,19 @@ public class Nrc implements Serializable {
         this.pk = nrcPK;
     }
 
-    public short getCupoDisponible() {
+    public Integer getCupoDisponible() {
         return cupoDisponible;
     }
 
-    public void setCupoDisponible(short cupoDisponible) {
+    public void setCupoDisponible(Integer cupoDisponible) {
         this.cupoDisponible = cupoDisponible;
     }
 
-    public short getCupoRegistrado() {
+    public Integer getCupoRegistrado() {
         return cupoRegistrado;
     }
 
-    public void setCupoRegistrado(short cupoRegistrado) {
+    public void setCupoRegistrado(Integer cupoRegistrado) {
         this.cupoRegistrado = cupoRegistrado;
     }
