+++ src/main/java/ec/edu/espe/arquitectura/escolastico/seguridad/model/Usuario.java
@@ -35,7 +35,7 @@ public class Usuario implements Serializable {
     @Temporal(TemporalType.DATE)
     private Date fechaCambioClave;
     @Column(name = "nro_intentos_fallidos", nullable = false)
-    private short nroIntentosFallidos;
+    private Integer nroIntentosFallidos;
     @Column(name = "fecha_ultima_sesion")
     @Temporal(TemporalType.TIMESTAMP)
     private Date fechaUltimaSesion;
@@ -130,11 +130,11 @@ public class Usuario implements Serializable {
         this.fechaCambioClave = fechaCambioClave;
     }
 
-    public short getNroIntentosFallidos() {
+    public Integer getNroIntentosFallidos() {
         return nroIntentosFallidos;
     }
 
-    public void setNroIntentosFallidos(short nroIntentosFallidos) {
+    public void setNroIntentosFallidos(Integer nroIntentosFallidos) {
         this.nroIntentosFallidos = nroIntentosFallidos;
     }
